Create Is This Sudoku's Solution Correct?
The Sudoku game spread out quickly all over the world; it became the most popular game around the planet. However, a lot of people disrespect the restrictions of the game by filling the matrix incorrectly.

Your job in this problem is to write a program that verifies if a filled matrix is either a solution or not for the given problem.

The matrix of the game is a 9 x 9 matrix of integers. In order to be a solution for the problem, each row and column must have their numbers from 1 to 9. Besides, if we divide the matrix in 9 regions of 3 x 3, each of these regions should also have the numbers from 1 to 9.

The example below shows a matrix as a solution for the problem.



1 3 2 | 5 7 9 | 4 6 8
4 9 8 | 2 6 1 | 3 7 5
7 5 6 | 3 8 4 | 2 1 9
------+-------+------
6 4 3 | 1 5 8 | 7 9 2
5 2 1 | 7 9 3 | 8 4 6
9 8 7 | 4 2 6 | 5 3 1
------+-------+------
2 1 4 | 9 3 5 | 6 8 7
3 6 5 | 8 1 7 | 9 2 4
8 7 9 | 6 4 2 | 1 5 3