Create Frog's game
The frog's game is very popular in Frogland.

It consists of a board with n areas, numbered from 0 to n-1.

As it is very popular, several players can play it. They alternate between rounds. At each round, the player number 1 rolls the dice first, and after the player 1, player 2 rolls the dice, and after that, player 3 rolls the dice and after that the n player rolls the dice ... and so on.

Initially the players place their frogs at area 0.

The areas are neutral, give a bonus or penalize the player. A neutral area has the value of 0, a positive area has a positive value and a negative area has a negative value.

Suppose that the value of the area is x.

If x is positive, it means the frog must jump forward (move forward) x areas. For example, if it's at the area number 4 and x is equal 3, then the frog must move to the area number 7.
If x is negative, it means the frog must jump backward (move backward) x areas.
If x is 0, then it means the frog doesn't have to move forward or backward.
So, as soon the dice is rolled, the frog jumps forward the number of areas shown by the dice and "falls" at an area. So it moves one more time according to the area it has fallen. After that, is the turn of the next player, even if the house gives a bonus or penalty.

One more important detail: the frogs can move a lot around the board. In other words, after the area numbered n-1, the next area is the area number 0. The same occurs in the opposite direction, for example, if the frog is at the area number 1 and has to move backward 4 areas, it must go to the area numbered n-3.

Your goal is: at the end of each round, identify in which area of the board the frogs are at.